refactor: CLI ロジックを xgrep_search::cli モジュールに公開
